+++ lib/improver/tests/wxcode/wxcode/test_wxcode_utilities.py
@@ -243,8 +243,10 @@ class Test_add_wxcode_metadata(IrisTest):
 
     def setUp(self):
         """Set up cube """
-        data = np.array([0, 1, 5, 11, 20, 5, 9, 10, 4,
-                         2, 0, 1, 29, 30, 1, 5, 6, 6]).reshape(2, 1, 3, 3)
+        data = np.array(
+            [0, 1, 5, 11, 20, 5, 9, 10, 4, 2, 0, 1, 29, 30, 1, 5, 6, 6],
+            dtype=np.float32
+        ).reshape(2, 1, 3, 3)
         self.cube = set_up_cube(data, 'air_temperature', 'K',
                                 realizations=np.array([0, 1]))
         self.wxcode = np.array(list(WX_DICT.keys()))
